move columns to AddOnFormSection.tsx
